% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_references.R
\name{add_new_references}
\alias{add_new_references}
\title{Append new references}
\usage{
add_new_references(complement, references)
}
\arguments{
\item{complement}{dataframe. References to be added.}

\item{references}{dataframe. References to which the references are added.}
}
\value{
A dataframe of references
}
\description{
Add imported references with unique keys to
the initial list of references.
}
\author{
Nicolas Mangin
}
