% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_row_table.R
\name{child_row_table}
\alias{child_row_table}
\title{Table with collapsible rows (3)}
\usage{
child_row_table(x, pos = NULL)
}
\arguments{
\item{x}{Data.frame. Table of references.}

\item{pos}{Integer. Position of the row.}
}
\value{
A data.table with collapsed rows.
}
\description{
Function formatting references with collapsible rows
}
\author{
Nicolas Mangin
}
