% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_references.R
\name{cite_references}
\alias{cite_references}
\title{Insert Citations in Text}
\usage{
cite_references()
}
\description{
Gadget for the selection and insertion of citations in Rmarkdown documents.
}
\author{
Nicolas Mangin
}
