% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{Clean and simplify strings}
\usage{
clean_string(x, simplify = FALSE)
}
\arguments{
\item{x}{Character. String which should be cleaned}

\item{simplify}{Logical. Whether only letters should be kept.}
}
\value{
A character string with only letters
}
\description{
Format authors' names so that they fit with bibtex specifications.
}
\author{
Nicolas Mangin
}
