% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Import References and Combine Them}
\usage{
combine()
}
\value{
An Excel file with the new complete reference list.
}
\description{
Application gathering references from different sources, identifying potential duplicates, and combining them in a single reference file.
}
\author{
Nicolas Mangin
}
