% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_references.R
\name{combine_references}
\alias{combine_references}
\title{Import References and Combine Them}
\usage{
combine_references()
}
\value{
An Excel file with the new complete reference list.
}
\description{
Application gathering references from different sources, identifying potential duplicates, and combining them in a signle reference file.
}
\author{
Nicolas Mangin
}
