% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_file.R
\name{create_reference_file}
\alias{create_reference_file}
\title{Create a reference file}
\usage{
create_reference_file()
}
\value{
A properly formatted reference.xlsx file in hte working directory.
}
\description{
Function to initialize a properly formated xlsx file to gather references.
}
\author{
Nicolas Mangin
}
