% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_references_server.R
\name{edit_references_server}
\alias{edit_references_server}
\title{Search and edit references}
\usage{
edit_references_server(id, course_paths, refdir)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{refdir}{Character. Path to the folder where the references are.}
}
\value{
Save edited list of references
}
\description{
Module facilitating their edition on a case by case basis.
}
\author{
Nicolas Mangin
}
