% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_references.R
\name{filter_references}
\alias{filter_references}
\title{Filter the reference based on input}
\usage{
filter_references(
  dataset,
  variable,
  filter_value = NULL,
  filter_type = "pattern"
)
}
\arguments{
\item{dataset}{Dataframe of tibble. References to filter.}

\item{variable}{Character string. Variable used for filtering.}

\item{filter_value}{Regex pattern, character string, numeric vector, or numeric value. Dependent on the the filter type}

\item{filter_type}{Character string. Specify whether the filter is a regex "pattern", a "selection" string, a numeric "range" or a numeric "value".}
}
\value{
A properly formatted string og references
}
\description{
Filter a dataset based on a filter value given by the user for a variable
according to the type of filter.
}
\author{
Nicolas Mangin
}
