% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_potential_duplicates.R
\name{find_potential_duplicates}
\alias{find_potential_duplicates}
\title{Identify potential duplicates in a list of references}
\usage{
find_potential_duplicates(x, distmethod = "qgram", maxdist = 10)
}
\arguments{
\item{x}{Tibble. Table with keys and titles.}

\item{distmethod}{Character. Method to compute distances between titles. Can be: osa, lv, dl, hammig, lcs, qgram, cosine, jaccard, jw, or soundex.}

\item{maxdist}{Numeric. Threshold to apply. Only titles the distance between which is smaller or equal to this number will be returned for check.}
}
\value{
A numeric vector with the row numbers of the potential duplicates
}
\description{
Function computing distances between titles and selecting those which are too close as defined by a maximum distance.
}
\author{
Nicolas Mangin
}
