% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_authors.R
\name{format_authors}
\alias{format_authors}
\title{Format Author Names}
\usage{
format_authors(author)
}
\arguments{
\item{author}{Character. String with names of authors.
Authors are separated by " and ".}
}
\value{
A character string with all the author names properly formatted.
}
\description{
Format authors' names so that they fit with bibtex specifications.
}
\author{
Nicolas Mangin
}
