% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citations.R
\name{format_citations}
\alias{format_citations}
\title{Format citations for insertion}
\usage{
format_citations(citations, pages)
}
\arguments{
\item{citations}{Character vector. Keys of the references to be inserted.}

\item{pages}{Character string. Pages if specified.}
}
\value{
A properly formatted string og references
}
\description{
Bring all citations together in a properly formatted string to be inserted.
}
\author{
Nicolas Mangin
}
