% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bib.R
\name{gen_bib}
\alias{gen_bib}
\title{Generate the Reference File}
\usage{
gen_bib(wdir = NULL, keys = NULL, journal_list = FALSE, author_list = FALSE)
}
\arguments{
\item{wdir}{Character string. Path to the desired working directory.}

\item{keys}{Character vector. List of bibtex keys to export as a bibtex database.}

\item{journal_list}{Logical. Whether a list of journals with the number of citations
should be produced.}

\item{author_list}{Logical. Whether a list of authors with the number of citations
should be produced.}
}
\value{
A bib file in the same folder as the Rmarkdown document.
}
\description{
Create the .bib reference file of the markdown document in which
the function is embedded and add it in the working directory.
Place this function in the setup of the markdown document.
}
\author{
Nicolas Mangin
}
