% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_references.R
\name{get_new_references}
\alias{get_new_references}
\title{Prepare New References for Combination}
\usage{
get_new_references(file)
}
\arguments{
\item{file}{character. Path to the .bib file.}
}
\value{
A list of references which can be appended to the main file.
}
\description{
Import and format references from bibtex files.
}
\author{
Nicolas Mangin
}
