% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_references.R
\name{get_references}
\alias{get_references}
\title{Load specified references files.}
\usage{
get_references(name, path)
}
\arguments{
\item{name}{Character. Name of the original file.}

\item{path}{Character. Path to the temporary file where the references to import are.}
}
\value{
References from one file.
}
\description{
Function applying the proper importation function and pre-cleaning the references to import.
}
\author{
Nicolas Mangin
}
