% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import References in bibliogR}
\usage{
import()
}
\value{
An updated database of references in the bibliogR package folder.
}
\description{
Gadget to import a list of references from Excel to R
}
\author{
Nicolas Mangin
}
