% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_citations.R
\name{insert_citations}
\alias{insert_citations}
\title{Insert Citations in Text}
\usage{
insert_citations()
}
\description{
Gadget for the selection and insertion of citations in Rmarkdown documents.
}
\author{
Nicolas Mangin
}
