% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journals.R
\docType{data}
\name{journals}
\alias{journals}
\title{journals}
\format{
List of collected journals with acronym, issn, and field.
\itemize{
\item  journal   character
\item  jnl   character
\item  issn   character
\item  field   character
}
}
\usage{
data(journals)
}
\description{
journals
}
\keyword{dataset}
