% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bib_entry.R
\name{make_bib_entry}
\alias{make_bib_entry}
\title{Write an entry for a bibtex file}
\usage{
make_bib_entry(
  bibtype,
  key,
  author,
  title,
  journal,
  year,
  month,
  volume,
  number,
  pages,
  publisher,
  booktitle,
  editor,
  institution,
  school,
  address,
  edition,
  note,
  doi,
  url,
  abstract,
  keywords,
  isbn,
  issn
)
}
\arguments{
\item{bibtype}{Character.}

\item{key}{Character.}

\item{author}{Character.}

\item{title}{Character.}

\item{journal}{Character.}

\item{year}{Character.}

\item{month}{Character.#'}

\item{volume}{Character.}

\item{number}{Character.}

\item{pages}{Character.}

\item{publisher}{Character.}

\item{booktitle}{Character.}

\item{editor}{Character.}

\item{institution}{Character.}

\item{school}{Character.}

\item{address}{Character.}

\item{edition}{Character.}

\item{note}{Character.}

\item{doi}{Character.}

\item{url}{Character.}

\item{abstract}{Character.}

\item{keywords}{Character.}

\item{isbn}{Character.}

\item{issn}{Character.}
}
\value{
A character vector in which each elements is a row of a bib entry.
}
\description{
Function taking one row of a reference list and writing the corresponding entry which will be added to a reference file.
}
\author{
Nicolas Mangin
}
