% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_key.R
\name{make_new_key}
\alias{make_new_key}
\title{Make a new bib key}
\usage{
make_new_key(name, year, keys)
}
\arguments{
\item{name}{Character. Author, institution, or school}

\item{year}{Year of publication.}

\item{keys}{Character vector. List of already used keys.}
}
\value{
A new bib key.
}
\description{
Function creating a new key for a reference.
}
\author{
Nicolas Mangin
}
