% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_references_server.R
\name{manage_references_server}
\alias{manage_references_server}
\title{Batch import new references or delete duplicates.}
\usage{
manage_references_server(id, references, refdir)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{references}{Reactive. Function containing a list of references.}

\item{refdir}{Character. Path to the folder where the references are.}
}
\description{
Module allowing the user to import multiple references from a file or to identify and remove duplicated references.
}
\author{
Nicolas Mangin
}
