% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_references_ui.R
\name{manage_references_ui}
\alias{manage_references_ui}
\title{Batch import new references or delete duplicates.}
\usage{
manage_references_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save edited list of references
}
\description{
Module allowing the user to import multiple references from a file or to identify and remove duplicated references
}
\author{
Nicolas Mangin
}
