% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_import.R
\name{modules_import}
\alias{modules_import}
\alias{import_ui}
\title{Module to import files}
\usage{
import_ui(id, label = "Select the file")
}
\arguments{
\item{id}{Character string. Id of the module}

\item{label}{Character string. Label on the interface}
}
\description{
Module to import files
}
