% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bib_import.R
\name{prepare_bib_import}
\alias{prepare_bib_import}
\title{Prepare bib files}
\usage{
prepare_bib_import(files, type = "article")
}
\arguments{
\item{files}{Character vector. List of paths to the .bib files to be imported.}

\item{type}{Bibtype to be imported. Defaut to "article".}
}
\value{
A clean list of references which can be imported with bibliogR interface.
}
\description{
Format .bib files to be imported in references.
}
\author{
Nicolas Mangin
}
