% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_new_references.R
\name{prepare_new_references}
\alias{prepare_new_references}
\title{Prepare New References for Combination}
\usage{
prepare_new_references(addref)
}
\arguments{
\item{addref}{Tibble. References list to clean.}
}
\value{
A list of references which can be appended to the main file.
}
\description{
Format references to allow their combination.
}
\author{
Nicolas Mangin
}
