% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ris_import.R
\name{prepare_ris_import}
\alias{prepare_ris_import}
\title{Prepare ris files}
\usage{
prepare_ris_import(files, type = "article")
}
\arguments{
\item{files}{Character vector. List of paths to the .ris files to be imported.}

\item{type}{Bibtype to be imported. Defaut to "article".}
}
\value{
A clean list of references which can be imported with bibliogR interface.
}
\description{
Format .ris files to be imported in references.
}
\author{
Nicolas Mangin
}
