% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_references_server.R
\name{search_references_server}
\alias{search_references_server}
\title{Filter a list of references}
\usage{
search_references_server(id, references)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{references}{Reactive. Function containing a list of references.}
}
\value{
Filtered list of references
}
\description{
Module allowing the user to find references using various filters.
}
\author{
Nicolas Mangin
}
