% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withchildrow.R
\name{withchildrow}
\alias{withchildrow}
\title{Create collapsible variables}
\source{
http://www.reigo.eu/2018/04/extending-dt-child-row-example/
}
\usage{
withchildrow(x, vars = NULL, opts = NULL, ...)
}
\arguments{
\item{x}{Tables to reformat.}

\item{vars}{Variables which should be collapsed.}

\item{opts}{Options.}

\item{...}{Other parameters}
}
\description{
Function to collapse variables in a table
}
