% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{callback}
\alias{callback}
\title{Table with collapsible rows (2)}
\usage{
callback(x, pos = NULL)
}
\arguments{
\item{x}{Data.frame. Table of references.}

\item{pos}{Integer. Position of the row.}
}
\value{
A data.table with collapsed rows.
}
\description{
Function formatting references with collapsible rows
}
\author{
Nicolas Mangin
}
