% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_references_ui.R
\name{edit_references_ui}
\alias{edit_references_ui}
\title{Search and edit references}
\usage{
edit_references_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save edited list of references
}
\description{
Module facilitating their edition on a case by case basis.
}
\author{
Nicolas Mangin
}
